refactor to chain pattern
