+++ service/src/main/java/uk/gov/hmcts/reform/fpl/model/cafcass/api/CafcassApiCaseDocument.java
@@ -1,0 +1,13 @@
+package uk.gov.hmcts.reform.fpl.model.cafcass.api;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class CafcassApiCaseDocument {
+    private String document_filename;
+    private boolean removed;
+    private String documentCategory;
+    private String documentId;
+}
